[MOVE] #21 - TabBar 별도의 파일로 분리
