+++ dnd-12th-2-iOS/dnd-12th-2-iOS/Presentation/Splash/SplashFeature.swift
@@ -20,7 +20,7 @@ struct SplashFeature {
         Reduce { state, action in
             switch action {
             case .loginCheck:
-                return .send(.loginCompleted(false))
+                return .send(.loginCompleted(true))
             default:
                 return .none
             }
